requirements.txt
praw==7.8.1
openai==1.99.9
python-dotenv==1.1.1
requests==2.32.3
